Switch to Git and define best practices for push access
The best practices include the new recommendation of doing Linux-style
signoffs that assert the Developer's Cerificate of Origin by implied
reference.

The DCO is committed separately on its own (detached) branch so that
we don't contaminate the package with its CC-by-sa/2.5 license.  I believe
that the DCO is merely aggregated with dctrl-tools in Git, and that this
does not make dctrl-tools derivative of the DCO.

Signed-off-by: Antti-Juhani Kaijanaho <antti-juhani@kaijanaho.fi>