msg.[ch] (and others): allow the message functions to handle format strings
There are some places in the code where it would simplify both the code
and the translations to have message strings be printf-formatted.  This
change changes the message function interfaces so that they can take
a variable number of arguments.  This required changes at all call sites.

At the same time, the key logic of the message functions is moved out of
the headers, leaving only the "shall we output this" test for inlining.

Signed-off-by: Antti-Juhani Kaijanaho <ajk@debian.org>