Add support for #-comments
Debian Policy allows, in debian/control, lines that start with the
'#'-character.  Those lines are treated as comments by Policy, and ignored.

This patch adds parser support for these comments in dctrl-tools.
Unfortunately, it is not possible (without major changes) to ignore
those comments semantically, since dctrl-tools represents paragraphs
and fields as substrings of the input file.

Accordingly, what this patch does is allow the parser to not choke on
comments.  Comments found inside a paragraph are left semantically
significant (ie. part of the field body in which they are found),
while comments otside of paragraphs are elided.

Signed-off-by: Antti-Juhani Kaijanaho <ajk@debian.org>