Partially revert fix for #525525 and add a migration path to it
The original fix for #525525 was a cold-turkey change in behavior.
This caused problems and drew objections.  Here, the idea is to
revert that change and add a migration path to it.

An option --ensure-dctrl is added, whose states semantics is to ensure that
the output format is in dctrl format.  It will be ignored if -n is also
given, as that deliberately asks for non-dctrl output.  In practical terms,
--ensure-dctrl makes the paragraph separator unconditional, as the cold
turkey change (in the absence of -n).  The idea is that people who prefer
the new behaviour can set up shell aliases or other similar mechanisms
that use --ensure-dctrl by default.

Another option, --compat is added to allow people who dislike the new
behaviour to make sure that they continue to receive it.  A --compat
option will countermand any previous --ensure-dctrl.  Thus a user can
put it in a shell alias, which will then continue to deliver the old
semantics even after --ensure-dctrl is made default.

The idea is that eventually --ensure-dctrl will become normal behaviour
and the switch becomes a no-op.

Signed-off-by: Antti-Juhani Kaijanaho <ajk@debian.org>